+++ pipecat-client-android/src/main/java/ai/pipecat/client/transport/TransportFactory.kt
@@ -1,8 +1,0 @@
-package ai.pipecat.client.transport
-
-/**
- * A creator of Transport objects.
- */
-fun interface TransportFactory {
-    fun createTransport(context: TransportContext): Transport
-}