+++ pipecat-client-android/src/main/java/ai/pipecat/client/types/Config.kt
@@ -1,5 +1,0 @@
-package ai.pipecat.client.types
-
-data class Config(
-    val config: List<ServiceConfig>
-)